Use normalized and deterministic build paths
Build systems can't be relied upon to be consistent in whether they
normalize paths, so we need to use a build path that is normalized in
the first place so there's no difference either way. This requires us
to sacrifice some brevity in order to minimise the chance of
name collisions.

Subdirs are now named with the full subport name plus an abbreviated
hash of the portpath. This means a collision is only possible when
different versions of the same subport have builds in progress at the
same time, and should be rare even then. If one does occur, statefile
locking should prevent any simultaneous use of a work dir, and the
Portfile hash will differ with high probability, so the consequences
will be the deletion of the build progress for the less recently built
version, and perhaps a confused developer.

See: https://trac.macports.org/ticket/72715