upgrade: add a progress counter
For each port to be upgraded or installed, prints a line just before
performing that operation, showing that port's position in the list and
the total number of ports in the list.

This required refactoring macports::upgrade to first generate a complete
list of operations to be performed, and then perform them afterwards,
so that the total number is known before the first one is performed.
This approach should hopefully enable other improvements in future.

In order to show an accurate total across what would previously have
been multiple calls to the upgrade proc, a new interface
macports::upgrade_multi is added, which allows upgrading multiple
"top-level" ports in one call.

Fixes: https://trac.macports.org/ticket/5001