mport{open,close}: avoid linear search of open_mports
Store as a dict indexed by porturl instead. It's still possible to open
the same port multiple times with different variations or options, so
the values are lists that are still searched linearly, but most of the
time they will be length 1 and they will very rarely contain more than
a handful of entries.

This also required changes to mportdepends because it was inspecting
open_mports. It's actually simpler now, maintaining its own mapping of
porturls to mports for each port opened.