Check archive availability asynchronously
Add an option to indicate to mportopen that the intent is to possibly
install the port, which causes it to run _eval_archive_available, which
queues a curl operation to check for the archive on a background thread
if needed. Use that option where appropriate.

Add a cancel operation to mport_fetch_thread. Use it if the operation
is still pending in a proc called from mportclose.

Avoid nested vwait calls in mport_fetch_thread code.

Add the option for _deptypes_for_target to return two lists of dep
types, those that are definitely needed and those that are needed only
if no archive is available.

Delay calling _archive_available as long as possible by processing deps
that are unconditionally needed first, thus getting more work done
while the background check may still be in progress.

Known issue: ctrl-c often won't interrupt dependency calculation.
Fixing this will at minimum require adding periodic Tcl_AsyncInvoke
calls to CurlGetSizeCmd (as currently done in CurlFetchCmd), which
requires using the curl multi API.