Add threaded file fetching
Before running appropriate targets, we queue asynchronous fetch or
archivefetch operations (depending on options, the target, and archive
availability) for the port and any dependencies that will be installed.
When the archivefetch or fetch targets are eventually run, they check
the results of these operations, waiting for them to complete first if
needed.

The number of simultaneous asynchronous fetch operations is controlled
by the new fetch_threads macports.conf option.

Progress is displayed when the currently executing target cannot
continue until a fetch operation completes.

Support for cancelling currently running asynchronous fetches has been
added, and this is done when mportclose is called and the port still
has one or more fetch operations outstanding.