Reduce use of pwd and cd
Change to a known good directory if pwd fails when port(1) is started.
Warn about this since it affects evaluation of the 'current' pseudoport
and file:// porturls with relative paths.

Remove use of a 'ports' directory in the current directory when no
sources are defined in sources.conf. This is not a good idea for
similar reasons to having . in PATH.

Closes: https://trac.macports.org/ticket/24687