portindex: don't use thread shared variables
Each of these is mutex protected and so has an added access cost, which
may be significant on older systems. We probably pay the cost of
getting the worker thread procedure's return value to the calling
thread regardless of whether we use it, so it makes sense to put all
the information there instead of in individual TSVs.

This also simplifies the accounting by removing the need to track job
numbers.